pyproject.toml
[tool.poetry]
name = "crowbotics-lite"
version = "0.1.0"
description = ""
authors = ["Hesbon5600 <hesbonkiptoo5600@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.7"}

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
